+++ src/kit/core/View.cpp
@@ -8,14 +8,6 @@ std::string View::getName() const {
     return name;
 }
 
-void View::contents() {
-//display stuff go here
-}
-
-void View::controls() {
-//controls go here
-}
-
 void View::setUi(Ui *ui) {
     View::ui = ui;
 }
@@ -26,6 +18,22 @@ void View::setViewsController(ViewsController *viewsController) {
 
 void View::setUtils(Utils *utils) {
     View::utils = utils;
+}
+
+void View::contents() {
+
+}
+
+void View::controls() {
+
+}
+
+void View::beforeEnter() {
+
+}
+
+void View::beforeExit() {
+
 }